requirements.txt
dvc==3.40.0
mlflow==2.8.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
flask==3.0.0
python-dotenv==1.0.0
pyyaml==6.0.1